IA/Assistente_Virtual/requirements.txt
gradio==3.24.1
google-generativeai==0.1.0